chore: fix lfs files that should have been pointers but aren't (#33413)
### Reason for this change

When checking out a fresh copy of the repo you are currently greeted with this message:

```
Encountered 32 file(s) that should have been pointers, but weren't:
  [list of the 32 files]
```

Additionally, when changing branches or updating branches some of these files frequently end up in a state that is confusing git, thus making it almost impossible to with the repo.

It appears that these files have not been migrated to lfs correctly.

### Description of changes

Run the following command to fix the files:

`git lfs migrate import --no-rewrite [list of the 32 files]`

### Describe any new or updated permissions being added

n/a

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*