chore(deps): update cdk-assets (#33415)
The lastest cdk-assets is required in cdk to mitigate a ECR upload issue. It includes the following fix: https://github.com/cdklabs/cdk-assets/pull/342. The following issues are related to this:

https://github.com/aws/aws-cdk/issues/30258
https://github.com/aws/aws-cdk/issues/31549
https://github.com/aws/aws-cdk/issues/33264

I am keeping #31549 open as it is still true. this [feature request](https://github.com/cdklabs/cdk-assets/issues/348) tracks the work to make cdk-assets compatible with containerd

Closes #30258 and closes #33264