chore(toolkit): tests don't exit
On GitHub Actions the `jest` test run doesn't exit because it is being
kept alive.

The reason is that we create a `CloudWatchLogMonitor` in one test, which
keeps on scheduling a timer with `setTimeout()` which keeps the Node
process alive.

I considered `unref()`fing the timer, but it seems a smaller and
therefore safer change to take the `CloudWatchLogMonitor` and deactivate
it.