chore(toolkit): tests don't exit (#33427)
On GitHub Actions the `jest` test run doesn't exit because it is being kept alive.

The reason is that we create a `CloudWatchLogMonitor` in one test, which keeps on scheduling a timer with `setTimeout()` which keeps the Node process alive. I'm not sure why this doesn't manifest on the CodeBuild build or a Mac laptop run, but it can be somewhat reproduced by running `npx jest --detectOpenHandles`, which *does* make the test hang on Mac as well.

I considered `unref()`fing the timer, but it seems a smaller and therefore safer change to take the `CloudWatchLogMonitor` and deactivate it.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*