Apply stack perm boundary to ref-time custom resources
The permissions boundary applied to a stack is not propagated to roles created
by custom resources where these enter the assembly at reference-resolution
time. Examples of this are cross-stack references and references to list values
which must be JSON stringified.

This change exposes the PB assigned to a stack and uses this to apply a custom
resource's host stack PB to its roles.