chore(cli-integ): make the version of `cli-lib-alpha` switchable
The `@aws-cdk/cli-lib-alpha` package is vended as an alpha package
in the old pipeline.

That means its version number is `2.178.1-alpha.0` for a real release,
or `2.178.1-alpha.999` for a release candidate, with the version
component matching the framework version.

In the new pipeline, we'll have to decouple from the framework version
to the CLI version anyway, and I don't want to have to rebuild this
version mangling logic in projen. Instead, we'll just give the library
the exact same version as the CLI.

Its alpha status is still obvious enough from its name, so I don't
think that will be a problem.