chore(cli-integ): make the version of `cli-lib-alpha` switchable (#33450)
The `@aws-cdk/cli-lib-alpha` package is vended as an alpha package in the old pipeline.

That means its version number is `2.178.1-alpha.0` for a real release, or `2.178.1-alpha.999` for a release candidate, with the version component matching the framework version.

In the new pipeline, we'll have to decouple from the framework version to the CLI version anyway, and I don't want to have to rebuild this version mangling logic in projen. Instead, we'll just always take the CLI version and append `-alpha.0`. That may mean we'll end up with a release candidate version of `-rc.0-alpha.0`, which looks weird but I don't think will break anything.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*