fix(core): wrong priority for tag aspects (#33460)
### ❗Important❗

This change is to fix behavior that was always wrong, starting in this [commit](https://github.com/aws/aws-cdk/pull/32097) released in CDK v2.172.0. In doing so, the order of your aspect execution may change. If you are inadvertently depending on an aspect ordering that was previously wrong (tagging was previously not prioritized as a mutating aspect), you could need to change your CDK code. We are not treating this as a breaking change because the previous order was always wrong.

### Reason for this change

Priority was not applied in https://github.com/aws/aws-cdk/pull/32333

### Description of changes

Fix missing priority

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*