fix(lambda-nodejs): do not require a frozen lockfile for bun (#32908)
### Issue #32906

Closes #32906.

### Reason for this change

When I implemented `bun` support, I accidentally used `--frozen-lockfile`, which caused issues when the lockfile contained additional entries other than what's being bundled in the docker container.

The issue has a small repro-case.

### Description of changes

I removed the `--frozen-lockfile` flag, which resolved the problem.

### Describe any new or updated permissions being added

N/A

### Description of how you validated changes

I added an additional package to the `bun.lockb` file in the integration test. This caused the issue to occur in the test suite. Once I made the changes in this PR, the test started passing again.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*