fix: pressing Ctrl-C when content is bundled leaves broken asset
When a bundling command is interrupted with Ctrl-C, the asset output
directory has already been created. On the next synthesis, we assume the
asset has already successfully been produced, don't do any bundling, and
upload it.

We will then have produced and uploaded a broken asset.

Instead, the common pattern to handle this is:

- Do the work into a temporary directory
- Rename the temporary directory to the target directory only if the
  work succeeds.